requirements.txt
amqp                    5.2.0
asgiref                 3.8.1
attrs                   24.2.0
billiard                4.2.1
celery                  5.4.0
certifi                 2024.8.30
cffi                    1.17.1
charset-normalizer      3.4.0
click                   8.1.7
click-didyoumean        0.3.1
click-plugins           1.1.1
click-repl              0.3.0
colorama                0.4.6
Django                  5.1.2
django-cors-headers     4.6.0
djangorestframework     3.15.2
h11                     0.14.0
idna                    3.10
kombu                   5.4.2
outcome                 1.3.0.post0
pip                     22.3.1
prompt_toolkit          3.0.48
pycparser               2.22
PySocks                 1.7.1
python-dateutil         2.9.0.post0
redis                   5.2.0
requests                2.32.3
selenium                4.25.0
setuptools              65.5.0
six                     1.16.0
sniffio                 1.3.1
sortedcontainers        2.4.0
sqlparse                0.5.1
trio                    0.27.0
trio-websocket          0.11.1
typing_extensions       4.12.2
tzdata                  2024.2
undetected-chromedriver 3.5.5
urllib3                 2.2.3
vine                    5.1.0
wcwidth                 0.2.13
wheel                   0.44.0
wsproto                 1.2.0